+++ aws-crt-kotlin/common/src/aws/sdk/kotlin/crt/auth/credentials/CachedCredentialsProvider.kt
@@ -18,7 +18,11 @@ package aws.sdk.kotlin.crt.auth.credentials
  * @return the newly-constructed credentials provider
  */
 public expect class CachedCredentialsProvider internal constructor(builder: CachedCredentialsProviderBuilder) : CredentialsProvider {
-    public companion object
+    public companion object { }
+
+    override fun close()
+    override suspend fun getCredentials(): Credentials
+    override suspend fun waitForShutdown()
 }
 
 public class CachedCredentialsProviderBuilder {
@@ -43,5 +47,4 @@ public class CachedCredentialsProviderBuilder {
 /**
  * Construct a new credentials provider using a builder
  */
-public fun CachedCredentialsProvider.Companion.build(block: CachedCredentialsProviderBuilder.() -> Unit):
-    CachedCredentialsProvider = CachedCredentialsProviderBuilder().apply(block).build()
+public fun CachedCredentialsProvider.Companion.build(block: CachedCredentialsProviderBuilder.() -> Unit): CachedCredentialsProvider = CachedCredentialsProviderBuilder().apply(block).build()