+++ aws-crt-kotlin/common/src/aws/sdk/kotlin/crt/auth/credentials/DefaultChainCredentialsProvider.kt
@@ -22,10 +22,13 @@ import aws.sdk.kotlin.crt.io.ClientBootstrap
  *
  * @return the newly-constructed credentials provider
  */
-public expect class DefaultChainCredentialsProvider internal constructor(builder: DefaultChainCredentialsProviderBuilder) :
-    CredentialsProvider {
+public expect class DefaultChainCredentialsProvider internal constructor(builder: DefaultChainCredentialsProviderBuilder) : CredentialsProvider {
 
-    public companion object
+    public companion object { }
+
+    override fun close()
+    override suspend fun getCredentials(): Credentials
+    override suspend fun waitForShutdown()
 }
 
 public class DefaultChainCredentialsProviderBuilder {
@@ -40,5 +43,4 @@ public class DefaultChainCredentialsProviderBuilder {
 /**
  * Construct a new credentials provider using a builder
  */
-public fun DefaultChainCredentialsProvider.Companion.build(block: DefaultChainCredentialsProviderBuilder.() -> Unit):
-    DefaultChainCredentialsProvider = DefaultChainCredentialsProviderBuilder().apply(block).build()
+public fun DefaultChainCredentialsProvider.Companion.build(block: DefaultChainCredentialsProviderBuilder.() -> Unit): DefaultChainCredentialsProvider = DefaultChainCredentialsProviderBuilder().apply(block).build()