+++ aws-crt-kotlin/common/src/aws/sdk/kotlin/crt/auth/credentials/ProfileCredentialsProvider.kt
@@ -12,8 +12,14 @@ import aws.sdk.kotlin.crt.io.TlsContext
  * A credentials provider that uses profile files.
  */
 public expect class ProfileCredentialsProvider
-internal constructor(builder: ProfileCredentialsProviderBuilder) : CredentialsProvider {
-    public companion object
+internal constructor(
+    builder: ProfileCredentialsProviderBuilder,
+) : CredentialsProvider {
+    public companion object { }
+
+    override fun close()
+    override suspend fun getCredentials(): Credentials
+    override suspend fun waitForShutdown()
 }
 
 public class ProfileCredentialsProviderBuilder {
@@ -50,5 +56,4 @@ public class ProfileCredentialsProviderBuilder {
 /**
  * Construct a new credentials provider using a builder.
  */
-public fun ProfileCredentialsProvider.Companion.build(block: ProfileCredentialsProviderBuilder.() -> Unit):
-    ProfileCredentialsProvider = ProfileCredentialsProviderBuilder().apply(block).build()
+public fun ProfileCredentialsProvider.Companion.build(block: ProfileCredentialsProviderBuilder.() -> Unit): ProfileCredentialsProvider = ProfileCredentialsProviderBuilder().apply(block).build()