+++ aws-crt-kotlin/common/src/aws/sdk/kotlin/crt/auth/credentials/StsAssumeRoleCredentialsProvider.kt
@@ -12,8 +12,14 @@ import aws.sdk.kotlin.crt.io.TlsContext
  * Provides credentials from STS by assuming a role
  */
 public expect class StsAssumeRoleCredentialsProvider
-internal constructor(builder: StsAssumeRoleCredentialsProviderBuilder) : CredentialsProvider {
-    public companion object
+internal constructor(
+    builder: StsAssumeRoleCredentialsProviderBuilder,
+) : CredentialsProvider {
+    public companion object { }
+
+    override fun close()
+    override suspend fun getCredentials(): Credentials
+    override suspend fun waitForShutdown()
 }
 
 public class StsAssumeRoleCredentialsProviderBuilder {
@@ -53,5 +59,4 @@ public class StsAssumeRoleCredentialsProviderBuilder {
 /**
  * Construct a new credentials provider using a builder.
  */
-public fun StsAssumeRoleCredentialsProvider.Companion.build(block: StsAssumeRoleCredentialsProviderBuilder.() -> Unit):
-    StsAssumeRoleCredentialsProvider = StsAssumeRoleCredentialsProviderBuilder().apply(block).build()
+public fun StsAssumeRoleCredentialsProvider.Companion.build(block: StsAssumeRoleCredentialsProviderBuilder.() -> Unit): StsAssumeRoleCredentialsProvider = StsAssumeRoleCredentialsProviderBuilder().apply(block).build()