+++ aws-crt-kotlin/common/src/aws/sdk/kotlin/crt/auth/credentials/StsWebIdentityCredentialsProvider.kt
@@ -13,8 +13,14 @@ import aws.sdk.kotlin.crt.io.TlsContext
  * authenticated in a mobile or web application with a web identity provider.
  */
 public expect class StsWebIdentityCredentialsProvider
-internal constructor(builder: StsWebIdentityCredentialsProviderBuilder) : CredentialsProvider {
-    public companion object
+internal constructor(
+    builder: StsWebIdentityCredentialsProviderBuilder,
+) : CredentialsProvider {
+    public companion object { }
+
+    override fun close()
+    override suspend fun getCredentials(): Credentials
+    override suspend fun waitForShutdown()
 }
 
 public class StsWebIdentityCredentialsProviderBuilder {
@@ -34,5 +40,4 @@ public class StsWebIdentityCredentialsProviderBuilder {
 /**
  * Construct a new credentials provider using a builder.
  */
-public fun StsWebIdentityCredentialsProvider.Companion.build(block: StsWebIdentityCredentialsProviderBuilder.() -> Unit):
-    StsWebIdentityCredentialsProvider = StsWebIdentityCredentialsProviderBuilder().apply(block).build()
+public fun StsWebIdentityCredentialsProvider.Companion.build(block: StsWebIdentityCredentialsProviderBuilder.() -> Unit): StsWebIdentityCredentialsProvider = StsWebIdentityCredentialsProviderBuilder().apply(block).build()