+++ aws-crt-kotlin/common/src/aws/sdk/kotlin/crt/io/HostResolver.kt
@@ -10,6 +10,10 @@ import aws.sdk.kotlin.crt.Closeable
 
 internal const val DEFAULT_MAX_ENTRIES = 8
 
-public expect class HostResolver(elg: EventLoopGroup, maxEntries: Int) : Closeable, AsyncShutdown {
+public expect class HostResolver(elg: EventLoopGroup, maxEntries: Int) :
+    Closeable,
+    AsyncShutdown {
     public constructor(elg: EventLoopGroup)
+    override fun close()
+    override suspend fun waitForShutdown()
 }