+++ aws-crt-kotlin/common/src/aws/sdk/kotlin/crt/io/TlsContext.kt
@@ -13,6 +13,7 @@ import aws.sdk.kotlin.crt.Closeable
  */
 public expect class TlsContext(options: TlsContextOptions? = null) : Closeable {
     public companion object {}
+    override fun close()
 }
 
 /**