+++ aws-crt-kotlin/jvm/src/aws/sdk/kotlin/crt/http/HttpClientConnectionManagerJVM.kt
@@ -20,7 +20,8 @@ import software.amazon.awssdk.crt.io.SocketOptions as SocketOptionsJni
 
 public actual class HttpClientConnectionManager actual constructor(
     public actual val options: HttpClientConnectionManagerOptions,
-) : Closeable, AsyncShutdown {
+) : Closeable,
+    AsyncShutdown {
 
     private val jniManager = HttpClientConnectionManagerJni.create(options.into())
 
@@ -52,11 +53,11 @@ public actual class HttpClientConnectionManager actual constructor(
         jniManager.releaseConnection(ktConn.jniConn)
     }
 
-    override fun close() {
+    actual override fun close() {
         jniManager.close()
     }
 
-    override suspend fun waitForShutdown() {
+    actual override suspend fun waitForShutdown() {
         jniManager.shutdownCompleteFuture.await()
     }
 }